# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/submissions/{submissionId}/grade":{"put":{"tags":["submission-controller"],"operationId":"gradeAssignment","parameters":[{"name":"submissionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"number","format":"double"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/enrollments/{studentId}/courses/{courseId}/lessons/complete":{"put":{"tags":["lesson-controller"],"operationId":"markLessonComplete","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"completedLessonsCount","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/enrollments/{enrollmentId}/drop":{"put":{"tags":["enrollment-controller"],"operationId":"dropCourse","parameters":[{"name":"enrollmentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/courses/{courseId}/instructor/{instructorId}":{"put":{"tags":["course-controller"],"operationId":"updateCourse","parameters":[{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"instructorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/submissions/student/{studentId}/assignment/{assignmentId}":{"post":{"tags":["submission-controller"],"operationId":"submitAssignment","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"assignmentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/students/{studentId}/quizzes/{quizId}/attempt":{"post":{"tags":["quiz-controller"],"operationId":"submitQuizAttempt","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"quizId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizAttemptRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuizAttempt"}}}}}}},"/api/enrollments/{enrollmentId}/certificate":{"post":{"tags":["enrollment-controller"],"operationId":"generateCertificate","parameters":[{"name":"enrollmentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/enrollments/student/{studentId}/course/{courseId}":{"post":{"tags":["enrollment-controller"],"operationId":"enrollStudent","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/courses/{courseId}/quizzes":{"post":{"tags":["quiz-controller"],"operationId":"createQuiz","parameters":[{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Quiz"}}}}}}},"/api/courses/{courseId}/lessons":{"post":{"tags":["lesson-controller"],"operationId":"addLesson","parameters":[{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Lesson"}}}}}}},"/api/courses/{courseId}/assignments":{"post":{"tags":["assignment-controller"],"operationId":"addAssignment","parameters":[{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignmentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Assignment"}}}}}}},"/api/courses/instructor/{instructorId}":{"get":{"tags":["course-controller"],"operationId":"getInstructorCourses","parameters":[{"name":"instructorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponse"}}}}}}},"post":{"tags":["course-controller"],"operationId":"createCourse","parameters":[{"name":"instructorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/enrollments/student/{studentId}":{"get":{"tags":["enrollment-controller"],"operationId":"getStudentEnrollments","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Enrollment"}}}}}}}},"/api/courses/{courseId}/students":{"get":{"tags":["course-controller"],"operationId":"getCourseStudents","parameters":[{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Student"}}}}}}}},"/api/courses/{courseId}/completion-rate":{"get":{"tags":["course-controller"],"operationId":"getCourseCompletionRate","parameters":[{"name":"courseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"number","format":"double"}}}}}}}},"/api/courses/search":{"get":{"tags":["course-controller"],"operationId":"searchCoursesByCategory","parameters":[{"name":"category","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponse"}}}}}}}},"/api/assignments/{assignmentId}/submissions":{"get":{"tags":["assignment-controller"],"operationId":"getAssignmentSubmissions","parameters":[{"name":"assignmentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Submission"}}}}}}}}},"components":{"schemas":{"CourseRequest":{"required":["category","duration","maxStudents","price","title"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"duration":{"type":"integer","format":"int32"},"price":{"type":"number"},"maxStudents":{"type":"integer","format":"int32"},"category":{"type":"string"}}},"QuizAttemptRequest":{"required":["answers","durationTakenMinutes"],"type":"object","properties":{"answers":{"type":"string"},"durationTakenMinutes":{"type":"integer","format":"int32"}}},"Course":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"instructor":{"$ref":"#/components/schemas/Instructor"},"duration":{"type":"integer","format":"int32"},"price":{"type":"number"},"maxStudents":{"type":"integer","format":"int32"},"category":{"type":"string"},"deleted":{"type":"boolean"}}},"Instructor":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ROLE_STUDENT","ROLE_INSTRUCTOR","ROLE_ADMIN"]},"expertise":{"type":"string"},"bio":{"type":"string"},"deleted":{"type":"boolean"}}},"Quiz":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"course":{"$ref":"#/components/schemas/Course"},"title":{"type":"string"},"questions":{"type":"string"},"passingScore":{"type":"number","format":"double"},"timeLimit":{"type":"integer","format":"int32"},"deleted":{"type":"boolean"}}},"QuizAttempt":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"quiz":{"$ref":"#/components/schemas/Quiz"},"student":{"$ref":"#/components/schemas/Student"},"score":{"type":"number","format":"double"},"status":{"type":"string"},"attemptNumber":{"type":"integer","format":"int32"},"submittedAt":{"type":"string","format":"date-time"},"studentAnswers":{"type":"string"},"deleted":{"type":"boolean"}}},"Student":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ROLE_STUDENT","ROLE_INSTRUCTOR","ROLE_ADMIN"]},"enrollmentDate":{"type":"string","format":"date"},"status":{"type":"string"},"deleted":{"type":"boolean"}}},"QuizRequest":{"required":["passingScore","questions","timeLimit","title"],"type":"object","properties":{"title":{"type":"string"},"questions":{"type":"string"},"passingScore":{"type":"number","format":"double"},"timeLimit":{"type":"integer","format":"int32"}}},"LessonRequest":{"required":["content","duration","title"],"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"duration":{"type":"integer","format":"int32"}}},"Lesson":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"course":{"$ref":"#/components/schemas/Course"},"title":{"type":"string"},"content":{"type":"string"},"orderIndex":{"type":"integer","format":"int32"},"duration":{"type":"integer","format":"int32"},"deleted":{"type":"boolean"}}},"AssignmentRequest":{"required":["description","dueDate","maxScore","title"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date-time"},"maxScore":{"type":"number","format":"double"}}},"Assignment":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"course":{"$ref":"#/components/schemas/Course"},"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date-time"},"maxScore":{"type":"number","format":"double"},"deleted":{"type":"boolean"}}},"CourseResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"instructorId":{"type":"integer","format":"int64"},"duration":{"type":"integer","format":"int32"},"price":{"type":"number"},"maxStudents":{"type":"integer","format":"int32"},"category":{"type":"string"}}},"RegisterRequest":{"required":["email","name","password"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"Enrollment":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"student":{"$ref":"#/components/schemas/Student"},"course":{"$ref":"#/components/schemas/Course"},"enrollmentDate":{"type":"string","format":"date"},"status":{"type":"string","enum":["ACTIVE","COMPLETED","DROPPED","EXPIRED"]},"progress":{"type":"number","format":"double"},"completionDate":{"type":"string","format":"date"},"deleted":{"type":"boolean"}}},"Submission":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"integer","format":"int64"},"assignment":{"$ref":"#/components/schemas/Assignment"},"student":{"$ref":"#/components/schemas/Student"},"content":{"type":"string"},"submittedAt":{"type":"string","format":"date-time"},"score":{"type":"number","format":"double"},"feedback":{"type":"string"},"deleted":{"type":"boolean"}}}}}}